Add APIs to libdmabufinfo to read sysfs DMA-BUF stats
The APIs will parse the DMA-BUF stats present at
/sys/kernel/dmabuf/buffers.

/sys/kernel/dmabuf/buffers/<inode_number> contains statistics about the
DMA-BUF with unique inode number <inode_number> such as its size,
exporter name and the number of times the buffer has been mmapped.

/sys/kernel/dmabuf/buffers/<inode_number>/attachments/<attachment_id>
contains stats about the attachment with ID <attachment_id> for DMA-BUF
of unique inode number <inode_number>.

/sys/kernel/dmabuf/buffers/<inode_number>/attachments/<attachment_id>/device
is a symlink that points to the importer device.
/sys/kernel/dmabuf/buffers/<inode_number>/attachments/<attachment_id>/map_counter
represents the number of distinct dev maps of the attachment.

Test: adb shell am bug-report,
Test: atest DmaBufSysfsStatsParser#TestReadDmaBufSysfsStats
Bug: 167709539
Change-Id: Id24eacd6205c096455152097aff6f6c9a699a3b0