procmeminfo: fix race in IsSmapsRollupSupported
smaps_rollup is falsely reported as unsupported in some cases. One
possible cause is that the process whose PSS is checked is killed so
/proc/<pid>/smaps_rollup is not present, which then causes libmeminfo
to assume that smaps_rollup is unsupported.

Instead, when checking for smaps_rollup support, check for the
existence of smaps_rollup belonging to the current process since that
process is by definition alive.

Bug: 181165549
Test: boots and collects memory
Change-Id: I08053c73de0168734938e408bd0c2993d8959ed3