procrank: add flag to filter descendants.
Figuring out whether a change to the zygote preloaded classes is
good or bad is difficult, because a change that reduces zygote's PSS by
a very noticeable megabyte can be bad if it increases that of its
children by a less noticeable 50 kilobytes. Add a flag to procrank that
lets us look at the PSS information for a subtree of the process
hierarchy, to make this easier.

Test: adb shell procrank -d $(adb shell pidof zygote64)
Change-Id: Icb24ad28e733cc63d635a9de4423644a0dd6a254