Split ProcessRecord into separate cpp/h files
This change moves the ProcessRecord declaration/definition into
separate processrecord.h/processrecord.cpp files in order to be able to
refer to ProcessRecords in the function signatures at smapinfo.h.

This change also removes the unique_ptr from procmem_ because it is not
needed, and makes ProcessRecord a class instead of a struct.

Test: Checked that showmap, librank, and procrank still work as
      expected.
Bug: 229147699
Change-Id: I501f89bb0daafd18d48e59b779fc69a18ca2eca5