Fixing failing tests in libmeminfo/libdmabufinfo
We currently have failing tests due to the transition from ION to
DMA-BUF heaps. We have tests that require to check `dev/ion`, which has
been replaced with `dev/dma_heap`, leading to failing tests. This CL aims to check that these tests are only running on devices which have not migrated to dma-buf heaps.

Test: Verified with cuttlefish that is running on dmabuf-heaps
Bug: 239105588
Change-Id: Ic2b99911703cd5c699a85e860fb49d7f1e8d8b7b