memevents: Migration from memhealth driver to BPF
Refactoring libmemevents to utilize BPF programs, and BPF ring-buffers,
to listen and collect data from kernel tracepoints. This removes the
need for the memhealth kernel driver.

BPF ring-buffers don't work with multiple consumers, meaning that every
consumer must have its own BPF ring buffer. Due to this limitation,
we have to predefine a ring-buffer(rb) for every service/client that
plans on creating a MemEventListener, and a BPF program(prog) for
every tracepoint each client is interested in. Therefore, the
MemEventListener class' constructor now requires a MemEventClient
parameter to map the listener to the corresponding client's BPF rb and
prog.

We are starting with only two clients, AMS and LMKD. AMS can only
register/listen for out-of-memory (OOM) kills. LMKD has the option to
register/listen to direct reclaim state changes (begin/end).

Test: Verified BPF ring buffers and progs are loaded
Bug: 315560026
Change-Id: I309cda49f8732558c9ff12f18e1cb8fdd8913748
Signed-off-by: Carlos Galo <carlosgalo@google.com>