Update Smaps() to populate swap_offsets_
This change adds an 'update_swap_usage' parameter to ReadVmaStats(),
which will allow for swap_offsets_ to be filled in for a given vma
without updating the usage field. Smaps() has been updated to fill in
swap_offsets_ so that libsmapinfo-backed tools can correctly calculate
swap values, allowing for Smaps() usage to be reinstated.

Test: - Check that `showmap <pid>` and `procrank -d <pid>` show the same
        values
      - Check that the swap totals in `procrank` are non-zero
Bug: 320480542
Change-Id: Ica99b09cba9e6ad629311c2b6676baccc369582e