Update VmaCallback return type
This change updates VmaCallback to return a bool value, and removes the
'const' qualifier on the input vma so that vmas can be modified from a
callback if needed. Existing instances of VmaCallback have been updated
to return a bool, with callers terminating early upon failure.

Callbacks in libmeminfo_test and libmeminfo_benchmark have been updated
to account for this.

Test: atest libmeminfo_test
      atest libmeminfo_benchmark (builds; fails for unrelated reason)
Bug: 320480542
Change-Id: Idc1a75ce110429082b8fc2c83cfcc692dfd31c40