BPF: remove duplicate cflags and update btf flags
With the addition of the -Wall, -Werror, -Wextra cflags to the bpf soong
module as default, they are no longer necessary in the bpf definition in
Android.bp files.

Additionally, 'btf: true' is now set by default by the build system. To counter this and maintain the existing behavior, add 'btf: false'.

Test: treehugger
Change-Id: I58688de93f5c0adea7e04e742562cf9fae981ed7
Signed-off-by: Neill Kapron <nkapron@google.com>